feat(config): make ticket reference optional
This commit introduces a new version of the configuration to make ticket
the ticket reference optional. When the `ticket` table is omitted, no
ticket is asked for. In the `ticket` table, a new `required` key is
added: when it is `true`, a ticket is asked for and required, as
previously. When it is `false`, a ticket is asked for, but the question
can be skipped.

Refs: #2