feat(command): ensure commands are run from a Git worktree
Until this commit, `git z commit`, `git z init` and `git z update` would
all three return an error if not run from withing a Git worktree.
However, the underlying error would differ depending on the command,
making it impossible to pretty-print properly.

In all cases, it was due to the impossibility to get the root of the
worktree, not the fact of not being in a worktree itself. This commit
fixes this by calling `ensure_in_git_worktree` before any other
operation. All three commands will then return the same error, which we
will be able to pretty-print.

Refs: #14