style(clippy): allow pattern_type_mismatch
The `clippy::pattern_type_mismatch` lint can conflict in occasions with
`clippy::needless_borrowed_reference`, which is warn by default.

In addition, the compiler diagnostics for some borrowing issues in
matches push for a solution that is conflicting with
`clippy::pattern_type_mismatch`.

As this seems not only to be non-standard style, but as well to be
confusing in regards to diagnostics, let’s keep
`clippy::needless_borrowed_reference` as a warning and disable
`clippy::pattern_type_mismatch`.

Refs: #13