build(deb): replace the script with a devshell command
The initial need for a script was to ensure the built package is
portable, by compiling it on a non-NixOS distribution. However, the
binary was still dynamically-linked.

By statically-linking the binary using the musl target, we can now build
a portable package even in a Nix environment. Let’s then remove the
dependency on Docker and make the script a command in the devshell
instead.

Refs: #29