build(script): make it behave properly outside of a Git repo
Before this commit, the build script was not checking the status of Git
commands. Therefore, if Git was available but the build was run outside
of a Git repo, like when installing via cargo from crates.io, a revision
was still “detected”, but empty. This was leading to the version being
reported as `git-z 0.2.0+` for instance.

With this commit, no revision information is shown when Git is not
available.