build(script): use `git describe --tags` to check the version
The build.rs includes the commit hash by default in the version, but
omits it when git-z is built exactly from a release tag. This is done by
comparing the current version with the output of `git describe`.
However, `git describe` works only with annotated tags, but the GitHub
action used for the release replaces the annotated tag with a
lightweight tag to the same revision.

Using `git describe --tags` should fix the issue.

Refs: #38